Migrate `taxPercent` fields to `BigDecimal` and introduce unpaginated project retrieval
- Updated `taxPercent` from `Double` to `BigDecimal` in tax-related entities and payloads for precision.
- Added an unpaginated endpoint to retrieve projects by organisation.
- Streamlined invoice processing with tax calculations and `Money` API for improved precision.
- Introduced `InvoiceTaxDetailResponse` and related enhancements to invoice handling.
- Updated database configuration for development environment.

+++ src/main/java/com/qbitspark/buildwisebackend/accounting_service/tax_mng/entity/TaxEntity.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -21,7 +22,7 @@ public class TaxEntity {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID taxId;
     private String taxName;
-    private double taxPercent;
+    private BigDecimal taxPercent;
     @Column(name = "tax_description", columnDefinition = "TEXT")
     private String taxDescription;
